refactor: move edgelessPageBlock slots to service
