+++ packages/blocks/src/page-block/edgeless/components/toolbar/connector/connector-tool-button.ts
@@ -76,7 +76,7 @@ export class EdgelessConnectorToolButton extends EdgelessToolButton<
     if (this._states.some(key => changedProperties.has(key))) {
       if (this._menu) {
         this.updateMenu();
-        this.edgeless.slots.edgelessToolUpdated.emit({
+        this.edgeless.tools.setEdgelessTool({
           type: this._type,
           mode: this.mode,
         });
