+++ packages/blocks/src/page-block/edgeless/components/toolbar/shape/shape-tool-button.ts
@@ -189,7 +189,7 @@ export class EdgelessShapeToolButton extends EdgelessToolButton<
     if (this._states.some(key => changedProperties.has(key))) {
       if (this._menu) {
         this.updateMenu();
-        this.edgeless.slots.edgelessToolUpdated.emit({
+        this.edgeless.tools.setEdgelessTool({
           type: this._type,
           shapeType: this.shapeType,
         });
